feat: support multi-phase multi-parent compilation for CJMP

Created-by: Dendy_Grobovshik
Commit-by: Dendy_Grobovshik
Merged-by: CangjiePL
Description: ## Change Details (Required)

Support multi-phase multi-parent compilation

## Change Type (Required)

**Simply save and click the checkbox, or when editing, change `[ ]` to `[x]` for the relevant item.**

- [x] Feature
- [ ] Bugfix
- [ ] Build Process or Auxiliary Tool Changes
- [ ] Documentation Update

## Self-Check of Changes (Required)

**Please do not modify or delete the following options. Simply save and click the checkbox, or when editing, change `[ ]` to `[x]` for the relevant item.**

### Local Compilation Verification Results:

- [x] Cangjie Compiler compiled successfully
- [ ] Cangjie Runtime compiled successfully
- [ ] Cangjie Standard Library compiled successfully
- [ ] Not applicable (select this option for auxiliary tool changes or documentation updates)

### Local Test Case Verification Results:

- [x] Pass. Please provide screenshots below.
- [ ] Not applicable. Please explain the reasons below.

Only excluded test failed, all other test cases passed.
![image.png](https://raw.gitcode.com/user-images/assets/7239667/3820e0b8-b7c1-468f-881d-94e314554352/image.png 'image.png')

## Related Issues (Required)

https://gitcode.com/Cangjie/cangjie_compiler/issues/149

See merge request: Cangjie/cangjie_compiler!895

+++ include/cangjie/Lex/Token.h
@@ -182,7 +182,7 @@ struct StringPart {
         this->end = end;
     }
 };
-using TokenVecMap = std::unordered_map<size_t, std::vector<Token>>;
+using TokenVecMap = std::unordered_map<unsigned int, std::vector<Token>>;
 const std::vector<TokenKind>& GetEscapeTokenKinds();
 } // namespace Cangjie
